Renamed SSO Client class
- added composer autoload
