+++ codesniffer.ruleset.xml
@@ -8,10 +8,13 @@
 
     <!-- Include the WordPress ruleset, with exclusions. -->
     <rule ref="WordPress">
+        <exclude name="Generic.Files.LowercasedFilename" />
+        <exclude name="WordPress.NamingConventions.ValidHookName" />
         <exclude name="Generic.WhiteSpace.ScopeIndent.IncorrectExact" />
         <exclude name="Generic.WhiteSpace.ScopeIndent.Incorrect" />
         <exclude name="PEAR.Functions.FunctionCallSignature.Indent" />
         <exclude name="WordPress.VIP.RestrictedFunctions" />
+        <exclude name="WordPress.VIP.SlowDBQuery" />
     </rule>
 
     <!-- Include sniffs for PHP cross-version compatibility. -->
