+++ composer.json
@@ -18,11 +18,27 @@
   "support": {
     "issues": "https://github.com/discourse/wp-discourse/issues"
   },
-  "require": {
-    "php": ">=5.3",
-    "composer/installers": "1.0.*"
-  },
-  "require-dev": {
-    "php-mock/php-mock-phpunit": "^0.3"
+  "autoload": {
+      "psr-4": {
+          "WPDiscourse\\": "lib"
+      },
+
+      "files" : [
+        "lib/utilities.php",
+        "lib/nonce.php",
+        "templates/html-templates.php",
+        "templates/template-functions.php",
+        "lib/discourse.php",
+        "lib/settings-validator.php",
+        "lib/admin.php",
+        "lib/sso.php",
+        "lib/wordpress-email-verification.php",
+        "lib/discourse-sso.php",
+        "lib/discourse-publish.php",
+        "lib/discourse-comment.php",
+        "lib/meta-box.php",
+        "lib/shortcodes/external-sso.php",
+        "lib/sso-login-form.php"
+      ]
   }
 }
